+++ semcore/feedback-form/CHANGELOG.md
@@ -6,7 +6,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 ### Added
 
-- `title` prop to `FeedbackRating`.
+- `notificationText` prop to `FeedbackRating`.
 
 ## [6.38.0] - 2024-12-18